Adding NLP work file
Can't get NLP to work in ProjectBot file. Error message of dictionary. When convert message to string before passing to processing function, it correctly identifies the pronoun but not the noun. 

For example:
msg = "I like pizza"
returns "You a ["
which is not the intended (right now intended is "You a pizza")